Switch to uint from int as quantities cannot be negative (#491)

+++ contracts/Pair.sol
@@ -597,7 +597,7 @@ contract Pair is
                 token,
                 sender,
                 receiver,
-                int256(tokenQty)
+                tokenQty
             );
             require(responseCode == HederaResponseCodes.SUCCESS, errorMessage);
         }
